adding multiple empty functional components for our routes, installing react-router defining routes, moving ingredients logic cu it's own route
