+++ package.json
@@ -31,6 +31,7 @@
     "react-dev-utils": "^12.0.1",
     "react-dom": "^18.1.0",
     "react-refresh": "^0.11.0",
+    "react-router-dom": "^6.3.0",
     "react-tooltip": "^4.2.21",
     "resolve": "^1.20.0",
     "styled-components": "^5.3.5"
