+++ src/components/Calendar/Calendar.module.scss
@@ -1,0 +1,4 @@
+.calendar {
+  display: flex;
+  justify-content: center;
+}
