+++ src/components/Fridge/Fridge.module.scss
@@ -1,0 +1,4 @@
+.fridge {
+  display: flex;
+  justify-content: center;
+}
