+++ src/components/LandingPage/LandingPage.module.scss
@@ -1,0 +1,4 @@
+.landing-page {
+  display: flex;
+  justify-content: center;
+}
