+++ src/components/Recipe/Recipe.module.scss
@@ -1,0 +1,4 @@
+.recipe {
+  display: flex;
+  justify-content: center;
+}
