+++ src/components/RecipeBook/RecipeBook.module.scss
@@ -1,0 +1,4 @@
+.recipe-book {
+  display: flex;
+  justify-content: center;
+}
