+++ src/components/RecipeBook/index.tsx
@@ -1,0 +1,16 @@
+import {
+  FC
+} from 'react';
+import styles from './RecipeBook.module.scss';
+
+export interface IRecipeBookProps {}
+
+const RecipeBook: FC<IRecipeBookProps> = (props) =>{
+  return (
+    <div className={styles['recipe-book']}>
+      RecipeBook
+    </div>
+  );
+};
+
+export default RecipeBook;
