+++ src/components/ShoppingCart/ShoppingCart.module.scss
@@ -1,0 +1,4 @@
+.shopping-cart {
+  display: flex;
+  justify-content: center;
+}
