+++ src/components/WeeklyPlan/WeeklyPlan.module.scss
@@ -1,0 +1,4 @@
+.weekly-plan {
+  display: flex;
+  justify-content: center;
+}
